fix: MobileVitalsContext update with new properties
